Move tracker class into snapshot module
